Add: main의 feed 컴포넌트화 - feed 데이터 mock 데이터로 분리 후 복수 피드 구현
