+++ src/components/Comment/Comment.scss
@@ -1,3 +1,5 @@
+@import '/src/styles/variables.scss';
+
 .feed__comment {
   margin-bottom: 5px;
