+++ src/components/Feed/Feed.scss
@@ -1,0 +1,127 @@
+@import '/src/styles/variables.scss';
+
+.feed {
+  background-color: white;
+  display: flex;
+  flex-direction: column;
+  border: 1px solid $grey-10;
+  border-radius: 2px;
+  margin-bottom: 15px;
+
+  .feed__header {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 15px 15px;
+
+    .feed__user {
+      display: flex;
+      align-items: center;
+
+      .feed_user-ID {
+        margin-left: 15px;
+      }
+    }
+  }
+
+  .feed__img {
+    width: 470px;
+    height: 470px;
+
+    img {
+      width: 100%;
+      height: 100%;
+    }
+  }
+
+  .feed__btns {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 8px 15px;
+
+    .feed__btns--box {
+      width: 120px;
+      display: flex;
+      justify-content: space-between;
+    }
+
+    i {
+      font-size: 25px;
+      color: rgba(0, 0, 0, 0.7);
+    }
+
+    #red-heart {
+      color: red;
+    }
+  }
+
+  .feed__likes {
+    display: flex;
+    align-items: center;
+    margin: 5px 0;
+    padding: 0 15px;
+
+    img {
+      width: 15px;
+      height: 15px;
+      border-radius: 50%;
+      margin-right: 5px;
+    }
+
+    span {
+      font-size: 13px;
+    }
+
+    .user-number {
+      font-weight: bold;
+    }
+  }
+
+  .feed__info {
+    display: flex;
+    align-items: center;
+    margin-bottom: 5px;
+    padding: 0 15px;
+
+    .feed__text {
+      font-size: 14px;
+      margin: 0 10px;
+    }
+
+    button {
+      color: $grey-50;
+      font-size: 13px;
+    }
+  }
+
+  .feed__comment-list {
+    padding: 0 15px;
+  }
+
+  .comment-form {
+    border-top: 1px solid $grey-10;
+
+    .comment__input {
+      width: 90%;
+      height: 40px;
+      outline: none;
+      border: none;
+      padding-left: 15px;
+    }
+
+    .submit-disabled {
+      border: none;
+      background-color: inherit;
+      color: blue;
+      opacity: 0.2;
+    }
+
+    .submit-abled {
+      border: none;
+      background-color: inherit;
+      color: blue;
+      opacity: 1;
+    }
+  }
+}
