+++ src/components/Footer/Footer.scss
@@ -1,4 +1,4 @@
-$grey-30: rgba(0, 0, 0, 0.3);
+@import '/src/styles/variables.scss';
 
 footer {
   color: $grey-30;
