+++ src/pages/minjaeJo/Login/Login.scss
@@ -1,8 +1,4 @@
-$grey-10: rgba(0, 0, 0, 0.1);
-$grey-20: rgba(0, 0, 0, 0.2);
-$grey-30: rgba(0, 0, 0, 0.3);
-$grey-40: rgba(0, 0, 0, 0.4);
-$grey-50: rgba(0, 0, 0, 0.5);
+@import '/src/styles/variables.scss';
 
 .login {
   width: 400px;
@@ -12,6 +8,9 @@ $grey-50: rgba(0, 0, 0, 0.5);
   flex-direction: column;
   align-items: center;
   border: 1px solid $grey-20;
+  margin-top: 200px;
+  margin-left: auto;
+  margin-right: auto;
 
   #title {
     font-family: Lobster;
