+++ src/styles/variables.scss
@@ -17,3 +17,11 @@ $color-dark-gr: #a8a8a8;
 $color-blue-dbt: #90cbf3;
 $color-dark-gr: 142;
 $color-gray-gr: 219;
+
+/*민재 scss 변수 */
+
+$grey-10: rgba(0, 0, 0, 0.1);
+$grey-20: rgba(0, 0, 0, 0.2);
+$grey-30: rgba(0, 0, 0, 0.3);
+$grey-40: rgba(0, 0, 0, 0.4);
+$grey-50: rgba(0, 0, 0, 0.5);
